requirements.txt
python-telegram-bot==20.6
pandas>=2.2.3
